Made CMIS binding configurable
