+++ drc_cmis/client/__init__.py
@@ -1,1 +1,2 @@
-from .client import CMISDRCClient  # noqa
+from .browser_client import CMISDRCClient  # noqa
+from .soap_client import SOAPCMISClient
