+++ tox.ini
@@ -25,7 +25,7 @@ deps =
   django30: Django~=3.0.0
   freezegun
 commands =
-  py.test tests/test_soap_client.py tests/test_soap_document.py \
+  py.test tests/ \
    --junitxml=reports/junit.xml \
    --cov=drc_cmis --cov-report xml:reports/coverage-{envname}.xml \
    {posargs}
